moved unit tests into auth_server, have yet to be optimized
